+++ SmurfSurvivors/core/src/com/smurfsurvivors/game/entity/Movable.java
@@ -1,5 +1,5 @@
 package com.smurfsurvivors.game.entity;
 
 public interface Movable {
-    void move(int dx, int dy);
+    void move(float dx, float dy);
 }
